Split system.py into focused utility modules
- Split system.py (325→110 lines) into process_utils.py and system_info.py
- Created process_utils.py (80 lines) for process management
- Created system_info.py (175 lines) for system information gathering
- Maintained compatibility wrapper in system.py
- Progress: 6 modules still exceed 300-line limit, continuing refactoring
